+++ libraries/kotlinx-metadata/jvm/ChangeLog.md
@@ -3,6 +3,7 @@
 ## Next
 
 - Add `JvmPropertyExtensionVisitor.visitSyntheticMethodForDelegate` for optimized delegated properties (KT-39055).
+- [`KT-48965`](https://youtrack.jetbrains.com/issue/KT-48965) Make the type of `KmValueParameter.type` non-null `KmType`
 
 ## 0.3.0