+++ libraries/kotlinx-metadata/jvm/ChangeLog.md
@@ -3,6 +3,10 @@
 ## Next
 
 - Add `JvmPropertyExtensionVisitor.visitSyntheticMethodForDelegate` for optimized delegated properties (KT-39055).
+- Add JVM-specific class flags:
+  - `JvmClassExtensionVisitor.visitJvmFlags`
+  - `JvmFlag.Class.HAS_METHOD_BODIES_IN_INTERFACE`
+  - `JvmFlag.Class.IS_COMPILED_IN_COMPATIBILITY_MODE`
 - [`KT-48965`](https://youtrack.jetbrains.com/issue/KT-48965) Make the type of `KmValueParameter.type` non-null `KmType`
 
 ## 0.3.0